Create removeElement.py
Difficulty: Easy 

A better solution than mine is a ptr j = 0 and a for loop from i to last index over the array that "sucks" the current value into j whenever the current position is NOT equal to the value and increments j when this happens. Afterwards return j which is a count of how many values were NOT the value to be removed, which is exactly the length of the "new array".